Added Linux 64 bit fixes for the sound system.
git-svn-id: svn://0x263a.de/tekuum_src@212 3ae3452d-76a7-4814-9b33-b51e13292e42

+++ sound/snd_system.cpp
@@ -325,7 +325,9 @@ void idSoundSystemLocal::Init() {
 	}
 
 	// make a 16 byte aligned finalMixBuffer
-	finalMixBuffer = (float *) ( ( ( (int)realAccum ) + 15 ) & ~15 );
+	// Techyon RB: 64 bit fix, changed int to intptr_t
+	finalMixBuffer = (float *) ( ( ( (intptr_t)realAccum ) + 15 ) & ~15 );
+	// Techyon END
 
 	graph = NULL;
